fix api

+++ src/redux/apis.js
@@ -12,7 +12,8 @@ request.interceptors.response.use(
   },
   (res) => {
     let response = res.response;
-    if (response.status === 401) {
+    if (response.url !== "/api/auth/login" && response.status === 401) {
+      debugger;
       // refresh page
       localStorage.clear();
       window.location.reload();
